Timestamp fixing and refactoring

+++ src/events/messageCreate/checkBanWords.js
@@ -9,28 +9,32 @@ const { Message } = require("discord.js");
  */
 
 module.exports = async (message) => {
-	const words = message.content.trim().toLowerCase().split(/ +/);
-	const cache = await itemsListCached(message.guildId);
-	const serverLocale = await checkGuildLocale(message.guildId);
-	const bannedWordLocales = {
-		"pt-BR": "Uma das palavras digitas na sua ultima mensagem está BANIDA desse servidor, evite usa-la!",
-		"en-US": "One of the words you typed in your last message is BANNED from this server, avoid using it!",
-	};
+	if (!message.author.bot) {
+		const words = message.content.trim().toLowerCase().split(/ +/);
+		const cache = await itemsListCached(message.guildId);
 
-	for (i of words) {
-		if (cache.includes(i)) {
-			await message.delete();
-			await message.author.send(bannedWordLocales[serverLocale]);
-			return;
+		const bannedWordLocales = {
+			"pt-BR": "Uma das palavras digitas na sua ultima mensagem está BANIDA desse servidor, evite usa-la!",
+			"en-US": "One of the words you typed in your last message is BANNED from this server, avoid using it!",
+		};
+
+		for (i of words) {
+			if (cache.includes(i)) {
+				const serverLocale = await checkGuildLocale(message.guildId);
+				await message.delete();
+				await message.author.send(bannedWordLocales[serverLocale]);
+				return;
+			}
 		}
-	}
-	const db_words = await GetBannedWords(message.guildId);
-	for (i of words) {
-		if (db_words.includes(i)) {
-			addListCache(message.guildId, i);
-			await message.delete();
-			await message.author.send(bannedWordLocales[serverLocale]);
-			return;
+		const db_words = await GetBannedWords(message.guildId);
+		for (i of words) {
+			if (db_words.includes(i)) {
+				addListCache(message.guildId, i);
+				const serverLocale = await checkGuildLocale(message.guildId);
+				await message.delete();
+				await message.author.send(bannedWordLocales[serverLocale]);
+				return;
+			}
 		}
 	}
 };
